# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ISP Billing API
  version: 1.0.0
servers:
  - url: https://your-host/api
paths:
...
  /audit-logs:
    get:
      summary: List tenant audit events with filters, pagination, and redacted payloads
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: actor
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 250 }
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
...
  /invites:
    get:
      summary: List pending tenant invites
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
    post:
      summary: Create a tenant invite for owner, admin, or operator access
      security:
        - bearerAuth: []
      responses:
        '201': { description: Created }
        '400': { description: Invalid payload }
        '403': { description: Insufficient role for requested invite role }
        '409': { description: User is already a member }
...
  /api-keys:
    get:
      summary: List tenant API keys without exposing secret material
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
    post:
      summary: Create a scoped tenant API key and return the plaintext key once
      security:
        - bearerAuth: []
      responses:
        '201': { description: Created }
        '400': { description: Invalid scopes or expiry }
        '403': { description: Owner or admin role required }
...
  /mikrotik/servers:
    get:
      summary: List MikroTik servers
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
    post:
      summary: Create MikroTik server
      security:
        - bearerAuth: []
      responses:
        '201': { description: Created }
...
  /invoices:
    get:
      summary: List tenant invoices with finance status and balances
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
...
  /portal-api/invoices:
    get:
      summary: List invoices for the authenticated customer portal account
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '401': { description: Invalid or expired token }
  /portal-api/payments:
    get:
      summary: List payments for the authenticated customer portal account
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '401': { description: Invalid or expired token }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY